Renames due to rebase of new resulsstore
